move to scalarfunction
