+++ datafusion/physical-expr/src/lib.rs
@@ -27,7 +27,9 @@ mod partitioning;
 mod physical_expr;
 pub mod planner;
 mod scalar_function;
-pub mod udf;
+pub mod udf {
+    pub use crate::scalar_function::create_physical_expr;
+}
 pub mod utils;
 pub mod window;
